test: add mocks to fix ci error
