+++ src/shared/i18n/__mocks__/index.ts
@@ -1,0 +1,5 @@
+import { TFunction } from 'i18next';
+
+export const t: TFunction = (key, defaultValue?, options?) => {
+  return key;
+};
