+++ test/setup.ts
@@ -1,39 +1,9 @@
 import _noop from 'lodash/noop';
-import _isNil from 'lodash/isNil';
 
-// 用于处理原生模块的mock
-jest.mock('../src/web/components/Image', () => 'mock-image');
-
-/**
- * mock request
- */
-const requestMockMapping: [string, any][] = [
-  [
-    '/info/website/info',
-    {
-      info: {
-        title: 'mock title',
-        content: 'mock content',
-      },
-    },
-  ],
-];
-jest.mock('@shared/utils/request', () => {
-  return (path: string) => {
-    const specMock = requestMockMapping.find((item) =>
-      path.startsWith(item[0])
-    );
-    if (!_isNil(specMock)) {
-      return Promise.resolve({
-        data: specMock[1],
-      });
-    }
-
-    return Promise.resolve({
-      data: null,
-    });
-  };
-});
+// 用于处理mock
+jest.mock('@web/components/Image', () => 'mock-image');
+jest.mock('@shared/i18n');
+jest.mock('@shared/utils/request');
 
 /**
  * mock console
